Fix incorrect use of the 'static' keyword.
Various globals provided by the SFEMP3Shield library were declared as static
in the header file.  This caused them to be re-instantiated in each
source file which included them.  The result was that from the
perspective of the sketch, these globals were all uninitialized and
did not work when used.

This patch corrects the globals to be declared extern in the header
file, and unadorned in the library source file.  This is standard
practice when exposing global variables in a library.